Do more work in filewriter.Writer rather than the children
